requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-documentai
google-cloud-storage
google-cloud-bigquery
